feat: :sparkles: First try to create a api for the config file
I created a class that manages and interact with the config file. Note that there's not a check to see if it is compliant to the format like for the samplesheet. I created some class for each parameter of the config, each of them has a getter and a setter (the setter interact with the Configuration file))