feat: :sparkles: Introduced live update of samplesheet from BCManager when batch is completed
Now each time a basecalling is successfully exucuted (or failed) the BCM updates immediatly the samplesheet. Note: the path to the files needs to be the one used by DGX 2: I passed the samplesheet as an env variable since they carry on in a sbatch. Easier than modify script.sh and then instructions.sh